requirements.txt
inflect==6.0.4
uvicorn==0.21.1
fastapi==0.95.1
transformers==4.34.1
peft==0.5.0
pandas==2.0.1
dirtyjson==1.0.8
diskcache==5.6.1
pymongo==4.5.0
loguru==0.7.2
python-dotenv==1.0.0
inflect
sentencepiece==0.1.99
protobuf==4.25.0
ruamel.yaml==0.18.6
jsonschema==4.21.1
pymongo==4.6.1